middelware

+++ routes/api.php
@@ -17,9 +17,10 @@ use Illuminate\Support\Facades\Route;
 */
 
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    return $request->user();
-});
+//Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
+//    return $request->user();
+//});
+
 
 
 Route::middleware('auth:sanctum')->get('/remove', function (Request $request) {
@@ -27,8 +28,9 @@ Route::middleware('auth:sanctum')->get('/remove', function (Request $request) {
     $user->tokens()->delete();
     return 'tokens are deleted';
 });
-
+//authcontroller
 Route::post('registerclient', 'AuthauController@registerclient');
+
 
 Route::post('registeragence', 'AuthauController@registeragence');
 
@@ -44,35 +46,45 @@ Route::get('/ala','HelloController@ala');
 
 
 //offer routes
-
-Route::post('/createoffer', 'OfferController@createoffer');
-
-
-Route::get('/getofferall', 'OfferController@getofferall');
-Route::get('/getofferid/{a}/{b}/{c}', 'OfferController@getofferid');
-Route::get('/getofferagence/{a}', 'OfferController@getofferagence');
-Route::get('/getoffercategorie/{a}', 'OfferController@getoffercategorie');
-
-Route::post('/updateoffer/{a}', 'OfferController@updateoffer');
-Route::delete('/deleteoffer/{a}', 'OfferController@deleteoffer');
+Route::middleware('auth:sanctum')->post('/createoffer', 'OfferController@createoffer');
+//Route::post('/createoffer', 'OfferController@createoffer');
+Route::middleware('auth:sanctum')->get('/getofferall', 'OfferController@getofferall');
+//Route::get('/getofferall', 'OfferController@getofferall');
+Route::middleware('auth:sanctum')->get('/getofferid/{a}/{b}/{c}', 'OfferController@getofferid');
+//Route::get('/getofferid/{a}/{b}/{c}', 'OfferController@getofferid}');
+Route::middleware('auth:sanctum')->get('/getofferagence/{a}', 'OfferController@getofferagence');
+//Route::get('/getofferagence/{a}', 'OfferController@getofferagence');
+Route::middleware('auth:sanctum')->get('/getoffercategorie/{a}', 'OfferController@getoffercategorie');
+//Route::get('/getoffercategorie/{a}', 'OfferController@getoffercategorie');
+Route::middleware('auth:sanctum')->post('/updateoffer/{a}', 'OfferController@updateoffer');
+//Route::post('/updateoffer/{a}', 'OfferController@updateoffer');
+Route::middleware('auth:sanctum')->delete('/deleteoffer/{a}', 'OfferController@deleteoffer');
+//Route::delete('/deleteoffer/{a}', 'OfferController@deleteoffer');
 
 //user routes
-Route::get('/getuser', 'UserController@getuser');
-Route::post('/updateagence/{id}', 'UserController@updateagence');
-Route::post('/updateclient/{id}', 'UserController@updateclient');
+Route::middleware('auth:sanctum')->get('/getuser', '\App\Http\Controllers\UserController@getuser');
+//Route::get('/getuser', 'UserController@getuser');
+Route::middleware('auth:sanctum')->post('/updateagence', '\App\Http\Controllers\UserController@updateagence');
+//Route::post('/updateagence/{id}', 'UserController@updateagence');
+Route::middleware('auth:sanctum')->post('/updateclient', '\App\Http\Controllers\UserController@updateclient');
+//Route::post('/updateclient/{id}', 'UserController@updateclient');
+//Route::middleware('auth:sanctum')->get('/searchagence', '\App\Http\Controllers\UserController@search');
 Route::get('/searchagence/{search}', 'UserController@search');
 
 
 // favorite routes
-
-Route::post('/addfavorite', 'FavoriteController@ajouterfavorit');
-Route::delete('/deletefavorite/{a}/{b}', 'FavoriteController@deletefavorite');
-Route::get('/getfavorite/{a}', 'FavoriteController@getfavorite');
+Route::middleware('auth:sanctum')->post('/addfavorite', 'FavoriteController@ajouterfavorit');
+//Route::post('/addfavorite', 'FavoriteController@ajouterfavorit');
+Route::middleware('auth:sanctum')->delete('/deletefavorite/{a}/{b}', 'FavoriteController@deletefavorite');
+//Route::delete('/deletefavorite/{a}/{b}', 'FavoriteController@deletefavorite');
+Route::middleware('auth:sanctum')->get('/getfavorite/{a}', 'FavoriteController@getfavorite');
+//Route::get('/getfavorite/{a}', 'FavoriteController@getfavorite');
 
 
 // messages routes
-
-
-Route::post('/addmessage', 'MessageController@ajoutermessage');
-Route::delete('/deletemessage/{a}', 'MessageController@deletemessage');
-Route::get('/getmessage/{a}', 'MessageController@getmessage');
+Route::middleware('auth:sanctum')->post('/addmessage', 'MessageController@ajoutermessage');
+//Route::post('/addmessage', 'MessageController@ajoutermessage');
+Route::middleware('auth:sanctum')->delete('/deletemessage/{a}', 'MessageController@deletemessage');
+//Route::delete('/deletemessage/{a}', 'MessageController@deletemessage');
+Route::middleware('auth:sanctum')->get('/getmessage/{a}', 'MessageController@getmessage');
+//Route::get('/getmessage/{a}', 'MessageController@getmessage');
